contour: only touch a native window handle under its own platform plugin
Fixes a segfault in the macOS test run, and the same latent crash on Windows.

CocoaWindowFrame reinterpret_casts QWindow::winId() to an NSView* and sends it
Objective-C messages. That is only true under the COCOA platform plugin -- under
offscreen, minimal or vnc, winId() is something else entirely, and messaging it
is a segmentation fault. Which is exactly what the GUI tests do: they run under
QT_QPA_PLATFORM=offscreen, so `toggleMaximized` crashed the moment the expose
hook reached the frame adapter.

Win32WindowFrame has the identical hazard -- handing a non-HWND to
SetWindowLongPtr is undefined -- and only got away with it because the Windows
test run happens not to reach it. Both are guarded now.

Both shadow backends already carried this guard; the frame adapters were written
later and did not, which is the whole of the bug.

Also fixes the clang-tidy findings in the test sources, which the earlier sweep
missed by only checking non-test files: a constexpr local in the wrong case
style, two #if defined that #ifdef says, and two casts to the type the
expression already had.

+++ src/contour/platform/WindowShadowTiles_test.cpp
@@ -45,7 +45,7 @@ TEST_CASE("renderWindowShadowTiles produces uploadable tiles", "[contour][shadow
                 for (auto const x: std::views::iota(0, image.width()))
                 {
                     auto const pixel = image.pixel(x, y);
-                    auto const alpha = static_cast<int>(qAlpha(pixel));
+                    auto const alpha = qAlpha(pixel);
                     if (qRed(pixel) > alpha || qGreen(pixel) > alpha || qBlue(pixel) > alpha)
                         ++violations;
                 }
@@ -65,7 +65,7 @@ TEST_CASE("renderWindowShadowTiles produces uploadable tiles", "[contour][shadow
         auto previous = -1;
         for (auto const x: std::views::iota(0, left.width()))
         {
-            auto const alpha = static_cast<int>(qAlpha(left.pixel(x, 0)));
+            auto const alpha = qAlpha(left.pixel(x, 0));
             INFO("x " << x);
             CHECK(alpha >= previous);
             previous = alpha;
